+++ validation/kubernetes/gatekeeper/main.tf
@@ -1,7 +1,8 @@
 terraform {}
 
 module "gatekeeper" {
-  source             = "../../../modules/kubernetes/gatekeeper"
-  cluster_id         = "foobar"
-  exclude_namespaces = []
+  source                         = "../../../modules/kubernetes/gatekeeper"
+  azure_service_operator_enabled = false
+  cluster_id                     = "foobar"
+  exclude_namespaces             = []
 }